+++ projects/_03_independent_cqc/_03_impute/fargate/impute_ind_cqc_ascwds_and_pir.py
@@ -5,6 +5,9 @@ import polars as pl
 from polars_utils import cleaning_utils as cUtils
 from polars_utils import utils
 from polars_utils.expressions import is_care_home
+from projects._03_independent_cqc._03_impute.fargate.utils.combine_ascwds_and_pir import (
+    merge_ascwds_and_pir_filled_post_submissions,
+)
 from projects._03_independent_cqc._03_impute.fargate.utils.convert_pir_people_to_filled_posts import (
     convert_pir_to_filled_posts,
 )
@@ -63,17 +66,16 @@ def main(cleaned_ind_cqc_source: str, destination: str) -> None:
 
     lf = convert_pir_to_filled_posts(lf)
 
-    # merge_ascwds_and_pir_filled_post_submissions
+    lf = merge_ascwds_and_pir_filled_post_submissions(lf)
 
-    # Uncomment this call when merge_ascwds_and_pir_filled_post_submissions is converted to polars.
-    # lf = model_imputation(
-    #     lf,
-    #     IndCQC.ascwds_pir_merged,
-    #     IndCQC.ascwds_rate_of_change_trendline_model,
-    #     IndCQC.imputed_filled_post_model,
-    #     care_home=False,
-    #     extrapolation_method="ratio",
-    # )
+    lf = model_imputation(
+        lf,
+        IndCQC.ascwds_pir_merged,
+        IndCQC.ascwds_rate_of_change_trendline_model,
+        IndCQC.imputed_filled_post_model,
+        care_home=False,
+        extrapolation_method="ratio",
+    )
 
     lf = model_imputation(
         lf,